feat(writer): support non-rectangular and non-aligned apertures
This commit takes advantage of the new feature in IDA ICE 5 to translate non-rectangular windows as well as rectangular windows that are not aligned with X and Y axis.

This commit removes the `frame-thickness` input which is not needed anymore as we don't try to merge the non-rectangular windows in the new version. This is technically a breaking change but since the previous commit was also a breaking change I keep this as a new feature.